Fixed multiples bugs

+++ ft_calloc.c
@@ -1,13 +1,24 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_calloc.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: thallard <thallard@student.42lyon.fr>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/11/23 01:52:37 by thallard          #+#    #+#             */
+/*   Updated: 2020/11/23 15:52:14 by thallard         ###   ########lyon.fr   */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-void    *ft_calloc(size_t len, size_t character)
+void	*ft_calloc(int len, int character)
 {
-    void    *ret;
+	void	*ret;
 
-    if (!(ret = malloc(len * character)))
-        return (NULL);
-    if (ret)  
-        ft_memset(ret, 0, len * character);
-    return (ret);
-
+	if (!(ret = malloc(len * character)))
+		return (NULL);
+	if (ret)
+		ft_memset(ret, 0, len * character);
+	return (ret);
 }
